Make LOD float, more lenient float parsing
Sampler's internal LOD should be a float since that is how it is used.
Also make float parsing more lenient (may accept uint or int) which
should make more inputs acceptable which were unnecessarily rejected
before.

+++ src/values/sampled-img.hpp
@@ -63,7 +63,7 @@ public:
         return new Struct(elements, names);
     }
 
-    const unsigned getImplicitLod() const {
+    const float getImplicitLod() const {
         return sampler.getImplicitLod();
     }
